feat(bookings): drop two step creation of new bookings
We did provide a two step work flow to create new bookings. The
intention was to simplify booking by providing a way to use booking
templates for quick booking. Providing booking templates is still a
goal, but the previous implementation was complicating basic bookings.

This patch drops the two steps to create bookings with a stock variant.